requirements.txt
aiofiles==0.5.0
appdirs==1.4.4
asgiref==3.2.10
astroid==2.0.4
ChatterBot==1.0.1
chatterbot-corpus==1.2.0
click==7.1.2
colorama==0.3.9
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.11.1
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
h11==0.9.0
importlib-metadata==1.7.0
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
joblib==0.16.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mathparse==0.1.2
mccabe==0.6.1
mysql-connector-python==8.0.21
nltk==3.5
numpy==1.17.4
openpyxl==3.0.2
packaging==20.4
pandas==0.25.3
Pillow==7.0.0
Pint==0.15
protobuf==3.13.0
psycopg2==2.8.5
pylint==2.1.1
pymongo==3.11.0
PyMySQL==0.9.3
pyparsing==2.4.7
pytesseract==0.3.2
python-dateutil==2.7.5
pytz==2019.3
PyYAML==3.13
regex==2020.7.14
selenium==3.141.0
six==1.11.0
SQLAlchemy==1.2.19
sqlparse==0.3.1
starlette==0.13.8
tqdm==4.48.2
urllib3==1.25.7
uvicorn==0.11.8
virtualenv==20.0.31
virtualenvwrapper-win==1.2.6
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.10.11
xlrd==1.2.0
zipp==3.1.0
